Move some variables off of stack

+++ components/calibration/include/calibration_emeter.h
@@ -9,8 +9,8 @@
 double snToFloat(uint32_t data, uint16_t radix);
 uint32_t floatToSn(double data, uint16_t radix);
 
-bool emeter_write(uint8_t reg, int registerValue);
-bool emeter_write_float(uint8_t reg, double value, int radix);
+bool emeter_write(uint8_t reg, uint32_t registerValue);
+bool emeter_write_float(uint8_t reg, double value, uint8_t radix);
 bool emeter_read(uint8_t reg, uint32_t *val);
 
 double emeter_get_fsv(void);
